% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_wm_study.R
\name{clear_wm_study}
\alias{clear_wm_study}
\title{Remove Areas & BindingConstraints before Weekly Margin Simulation}
\usage{
clear_wm_study(opts = simOptions())
}
\arguments{
\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
An updated list containing various information about the simulation.
}
\description{
Remove Areas & BindingConstraints before Weekly Margin Simulation
}
\examples{
\dontrun{

opts <- setSimulationPath(path = "path/to/simulation/", simulation = "input")
clear_wm_study(opts)

}
}
