% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_all_margins.R
\name{compute_all_margins}
\alias{compute_all_margins}
\title{Compute upward or downward margins}
\usage{
compute_all_margins(date, mcYear, exclude = c("lu_be", "lu_de"),
  margin = c("upward", "downward"), virtual_areas = c("lac", "pump_d",
  "turb_d", "pump_w", "turb_w"), opts = antaresRead::simOptions())
}
\arguments{
\item{date}{Date of the study.}

\item{mcYear}{An MC year.}

\item{exclude}{Areas to exclude.}

\item{margin}{Type of margin to compute, \code{upward} or \code{downward}, can be abbreviated.}

\item{virtual_areas}{List of virtuals areas.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
a list of \code{data.table} with 2 slots: \code{areas} and \code{links}.
}
\description{
From Antares results, calculate upward or downward margins for all areas given a MC year.
}
\examples{
\dontrun{

# todo

}
}
