% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_margins.R
\name{compute_margins}
\alias{compute_margins}
\title{Compute upward or downward margins}
\usage{
compute_margins(date, area = "fr", margin = c("upward", "downward"),
  formula_cnes = TRUE, mcYears = "all", virtual_areas = c("lac",
  "pump_d", "turb_d", "pump_w", "turb_w"),
  opts = antaresRead::simOptions())
}
\arguments{
\item{date}{Date of the study.}

\item{area}{Area studied.}

\item{margin}{Type of margin to compute, \code{upward} or \code{downward}, can be abbreviated.}

\item{formula_cnes}{For downward margins, use CNES formula or not.}

\item{mcYears}{Index of the Monte-Carlo years to import, by default all of them. Passed to \code{\link[antaresRead]{readAntares}}.}

\item{virtual_areas}{List of virtuals areas.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
a list of \code{data.table} with 3 slots: \code{data_area},
 \code{margin_area_solo} and \code{margin_area_inter}.
}
\description{
From Antares results, calculate upward or downward margins for an area.
}
\examples{
\dontrun{

library(antaresRead)
library(antaresWeeklyMargin)

opts <- setSimulationPath("path/to/study/")

# Upward margins for France
up <- compute_margins(
  date = "2018-06-23",
  area = "fr",
  margin = "upward"
)

# Downward margins
# can be slow - all clusters are read
down <- compute_margins(
  date = "2018-06-23", 
  area = "fr",
  margin = "downward"
)

}
}
