% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mono.R
\name{compute_mono}
\alias{compute_mono}
\title{Compute Monotone Data}
\usage{
compute_mono(start = "2016-11-05", date = "2016-11-10 17:00:00",
  area = "fr", mcYears = "all", opts = antaresRead::simOptions())
}
\arguments{
\item{start}{Starting date}

\item{date}{Date desired for the monotone.}

\item{area}{Area studied.}

\item{mcYears}{Index of the Monte-Carlo years to import, by default all of them. Passed to \code{\link[antaresRead]{readAntares}}.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
a list of \code{data.table}
}
\description{
Compute Monotone Data
}
\examples{
\dontrun{

# todo

}
}
