% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_markdown.R
\name{create_markdown}
\alias{create_markdown}
\title{Create directory with SMTA markdown}
\usage{
create_markdown(path, format = c("report", "slides"),
  date_start = NULL, date_margins = NULL, date_mono = NULL,
  opts = antaresRead::simOptions())
}
\arguments{
\item{path}{Path to a directory, if directory doesn't exist, it will be created.}

\item{format}{Format of markdown output : report or slides.}

\item{date_start}{Study start date.}

\item{date_margins}{Date for margins computation.}

\item{date_mono}{Date-time for representing monotone.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\description{
Create directory with SMTA markdown
}
\examples{
\dontrun{

create_markdown("SMTA_rapport", "report")

}
}
