% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wm_cluster.R
\name{create_wm_cluster}
\alias{create_wm_cluster}
\title{Create clusters for Weekly Margins simulation}
\usage{
create_wm_cluster(data, opts = antaresRead::simOptions())
}
\arguments{
\item{data}{a \code{data.table} obtained from \code{\link{read_planning}}.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\description{
Create clusters for Weekly Margins simulation
}
\examples{
\dontrun{

# TODO

}
}
