% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wm_hydro_areas.R
\name{create_wm_hydro_areas}
\alias{create_wm_hydro_areas}
\title{Create Others Areas Hydro setup for Weekly Margins simulation}
\usage{
create_wm_hydro_areas(start, opts = antaresRead::simOptions())
}
\arguments{
\item{start}{Starting date of simulation.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\description{
Create Others Areas Hydro setup for Weekly Margins simulation
}
\examples{
\dontrun{

# todo

}
}
