% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wm_load_fr.R
\name{create_wm_load_fr}
\alias{create_wm_load_fr}
\title{Create Load TS for France}
\usage{
create_wm_load_fr(path, start, start_prev_hebdo, type = c("prevu", "premis",
  "offset"), opts = antaresRead::simOptions())
}
\arguments{
\item{path}{Path to CNES directory, see \code{\link{read_cnes}} for usage.}

\item{start}{Beginning of the simulation.}

\item{start_prev_hebdo}{Beginning of previsions.}

\item{type}{Forecast to use \code{prevu} or \code{premis}.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\description{
Create Load TS for France
}
\examples{
\dontrun{

create_wm_load_fr("Prev_hebdo_CNES/", "2018-02-27", "prevu")

}
}
