% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wm_misc.R
\name{create_wm_misc}
\alias{create_wm_misc}
\title{Create Misc-gen for Weekly Margins simulation}
\usage{
create_wm_misc(data, start = NULL, opts = antaresRead::simOptions())
}
\arguments{
\item{data}{a \code{data.table} obtained from \code{\link{read_forfait_oa}}.}

\item{start}{If specified, data will be filtered from given date to 7 days after.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\description{
Create Misc-gen for Weekly Margins simulation
}
\examples{
\dontrun{

# set path to your simulation
opts <- setSimulationPath(path = "path/to/simulation/", simulation = "input")

# Read OA files
oa <- read_forfait_oa(path = "path/to/hydro_forfait/")

# Create ROR series in Antares
create_wm_misc(data = oa, start = "2018-01-04")

}
}
