% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wm_ntc.R
\name{create_wm_ntc}
\alias{create_wm_ntc}
\title{Create Links for Weekly Margins simulation}
\usage{
create_wm_ntc(data, start = NULL, startday = "samedi", sort_links = TRUE,
  opts = antaresRead::simOptions())
}
\arguments{
\item{data}{a \code{data.table} containing NTC datas.}

\item{start}{Starting day of the simulation, data between previous \code{startday}
and next \code{startday}-1, by default between \code{samedi} and \code{vendredi}.}

\item{startday}{Day of week to start simulation.}

\item{sort_links}{Reorder other links to match the desired week.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\description{
Create Links for Weekly Margins simulation
}
\examples{
\dontrun{

# set path to your simulation
opts <- setSimulationPath(path = "path/to/simulation/", simulation = "input")

# NTC data
ntc <- fread("path/to/data.csv")

# Create links in Antares
create_wm_ntc(ntc, start = "2018-01-04")

}
}
