% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wm_ntc_tp.R
\name{create_wm_ntc_tp}
\alias{create_wm_ntc_tp}
\title{Create Links for de/nl/be for Weekly Margins simulation}
\usage{
create_wm_ntc_tp(data, start, default = 1468,
  opts = antaresRead::simOptions())
}
\arguments{
\item{data}{a \code{data.table} containing NTC datas.}

\item{start}{Starting day of the simulation.}

\item{default}{Default value to use if \code{forecasttransfercapacity} is 0.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\description{
Create Links for de/nl/be for Weekly Margins simulation
}
\examples{
\dontrun{

opts <- setSimulationPath(path = "test_case/", simulation = "input")

ntc_tp <- read_ntc(path = "inputs/ntc_be_de_nl/")

create_wm_ntc_tp(data = ntc_tp, start = "2018-02-24", opts = opts)

}
}
