% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wm_ror.R
\name{create_wm_ror}
\alias{create_wm_ror}
\title{Create ROR for Weekly Margins simulation}
\usage{
create_wm_ror(data, start = NULL, startday = "samedi",
  opts = antaresRead::simOptions())
}
\arguments{
\item{data}{a \code{data.table} obtained from \code{\link{read_forfait_oa}}.}

\item{start}{If specified, data will be filtered from given date to 7 days after.}

\item{startday}{Day of week to start simulation.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\description{
Create ROR for Weekly Margins simulation
}
\examples{
\dontrun{

# todo

}
}
