% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wm_ts.R
\name{create_wm_ts}
\alias{create_wm_ts}
\title{Create load, wind & soloar time series for Weekly Margins simulation}
\usage{
create_wm_ts(data, start = NULL, opts = antaresRead::simOptions())
}
\arguments{
\item{data}{a \code{data.table} obtained from \code{\link{read_meteologica2}}.}

\item{start}{If specified, data will be filtered from given date to 7 days after.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\description{
Create load, wind & soloar time series for Weekly Margins simulation
}
\examples{
\dontrun{

# set path to your simulation
opts <- setSimulationPath(path = "path/to/simulation/", simulation = "input")


# Meteologica forecast
formet <- read_meteologica2(path = "path/to/PrevisionMeteologica/Europe")

# Create time series
create_wm_ts(data = formet, start = "2018-04-01", opts = opts)

}
}
