% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wm_fr_ts.R
\name{create_wm_ts_fr}
\alias{create_wm_ts_fr}
\title{Create Solar and Wind TS for France}
\usage{
create_wm_ts_fr(data, start, opts = antaresRead::simOptions())
}
\arguments{
\item{data}{a \code{data.table} obtained from \code{\link{read_forfait_oa}}.}

\item{start}{Beginning of the simulation.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\description{
Create Solar and Wind TS for France
}
\examples{
\dontrun{

opts <- setSimulationPath(path = "test_case/", simulation = "input")

oa <- read_forfait_oa(path = "inputs/OA/")

create_wm_ts_fr(data = oa, start = "2018-01-02")

}
}
