% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-margins.R
\name{draw_margins}
\alias{draw_margins}
\title{Draw remaining capacity line chart}
\usage{
draw_margins(data_margin, area = "fr", nb_MC = ncol(data_margin) - 1,
  num_week = NULL)
}
\arguments{
\item{data_margin}{a \code{data.table}, first column must be the datetime,
the others results of Monte-Carlo simulations.}

\item{area}{Name of the area, used in chart title.}

\item{nb_MC}{Number of Monte-Carlo years to draw.}

\item{num_week}{Week number to be displayed in chart title.}
}
\value{
a \code{dygraphs} htmlwidget.
}
\description{
Draw remaining capacity line chart
}
\examples{
\dontrun{

# TODO

}
}
