% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-mono.R
\name{draw_mono}
\alias{draw_mono}
\title{Draw a Monotone}
\usage{
draw_mono(data, area = "auto", main = "auto", label = "auto")
}
\arguments{
\item{data}{A \code{data.table} with at leat two columns.}

\item{main}{Main plot title (optional).}

\item{label}{Serie's label (optional).}
}
\value{
a dygrgaph htmlwidget
}
\description{
Draw a Monotone
}
\examples{

dat <- data.frame(
  x = seq_len(100),
  y = sort(rnorm(100, 10, 5))
)
draw_mono(dat)

}
