% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-prod-MC.R
\name{draw_prod_MC}
\alias{draw_prod_MC}
\title{Draw Stack Production for a MC Year}
\usage{
draw_prod_MC(data, area = NULL, mc_year = NULL, date_i = NULL)
}
\arguments{
\item{data}{An object of class \code{antaresDataTable}.}

\item{area}{Area for which the graph is drawn}

\item{mc_year}{Monte-Carlo year to use.}

\item{date_i}{A new initial date to replace the original date in the data.}
}
\description{
Draw Stack Production for a MC Year
}
\examples{
\dontrun{

draw_prod_MC(data, "fr", 645)

}
}
