% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-scenario.R
\name{draw_scenario}
\alias{draw_scenario}
\title{Draw Exchange scenario}
\usage{
draw_scenario(data, var_value = "somme", var_date = "time",
  grid_size = 100)
}
\arguments{
\item{data}{A \code{data.table}.}

\item{var_value}{Name of variable with values to represent.}

\item{var_date}{Name of variable with datetime.}

\item{grid_size}{Size of discretization grid.}
}
\description{
Draw Exchange scenario
}
