% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-series.R
\name{draw_series}
\alias{draw_series}
\title{Draw a time series}
\usage{
draw_series(data, serie = "LOAD", mcYears = 1:51, main = serie)
}
\arguments{
\item{data}{Data to visualize, an \code{antaresDataTable} object.}

\item{serie}{Name of the serie.}

\item{mcYears}{Monte-Carlo years to represent.!}

\item{main}{Title for the chart.}
}
\value{
a \code{dygraphs} object.
}
\description{
Draw a time series
}
\examples{
\dontrun{

# TODO

}
}
