% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-stack-hist.R
\name{draw_stack_hist}
\alias{draw_stack_hist}
\title{Draw Remaining Capacity}
\usage{
draw_stack_hist(marge_seule, marge_inter, area = NULL)
}
\arguments{
\item{marge_seule}{Margin data}

\item{marge_inter}{Margin data}

\item{area}{Concerned area, used in title.}
}
\description{
Draw Remaining Capacity
}
\examples{
\dontrun{

draw_stack_hist(marge_seul_fr, marge_inter_fr, "fr")

}
}
