% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-upward-margin.R
\name{draw_upward_margin}
\alias{draw_upward_margin}
\title{Draw remaining capacity line chart}
\usage{
draw_upward_margin(upward_margin, area = "fr", type = c("inter", "seul"),
  nb_MC = ncol(upward_margin) - 1, num_week = NULL)
}
\arguments{
\item{upward_margin}{a \code{data.table}, first column must be the datetime,
the others results of Monte-Carlo simulations.}

\item{area}{Name of the area, used in chart title.}

\item{type}{Initial or final remaining capacity.}

\item{nb_MC}{Number of Monte-Carlo years to draw.}

\item{num_week}{Week number to be displayed in chart title.}
}
\value{
a \code{dygraphs} htmlwidget.
}
\description{
Draw remaining capacity line chart
}
\examples{
\dontrun{

# TODO

}
}
