% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_wm.R
\name{flow_based_wm}
\alias{flow_based_wm}
\title{Flow Based setup for Weekly Margin Simulation}
\usage{
flow_based_wm(path_data, start_prev_hebdo,
  opts = antaresRead::simOptions())
}
\arguments{
\item{path_data}{Path to the folder containing data used for Flow Based.}

\item{start_prev_hebdo}{Date of forecasts, with format \code{\%Y-\%m-\%d}.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}.}
}
\description{
Flow Based setup for Weekly Margin Simulation
}
