% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rte-api.R
\name{get_eco2mix}
\alias{get_eco2mix}
\title{Get data from eco2mix via Open-RTE API}
\usage{
get_eco2mix(from = NULL, to = NULL, user = NULL, proxy_pwd = NULL)
}
\arguments{
\item{from}{date from which to retrieve data.}

\item{to}{date until which to recover data.}

\item{user}{Username (NNI) for proxy if needed.}

\item{proxy_pwd}{Password for proxy if needed.}
}
\value{
a \code{data.table}
}
\description{
Get data from eco2mix via Open-RTE API
}
\examples{
\dontrun{

# TODO

}
}
