% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rte-api.R
\name{get_hydraulique_fil_de_l_eau_eclusee}
\alias{get_hydraulique_fil_de_l_eau_eclusee}
\title{Get hydraulique data drom eco2mix}
\usage{
get_hydraulique_fil_de_l_eau_eclusee(from = NULL, to = NULL,
  resource = "tr", user = NULL, proxy_pwd = NULL)
}
\arguments{
\item{from}{date from which to retrieve data, if \code{NULL} set to previous saturday before previous friday.}

\item{to}{date until which to recover data, if \code{NULL} set to previous friday.}

\item{resource}{resource to use between real time ("tr") or consolidated data ("cons").}

\item{user}{Username (NNI) for proxy if needed.}

\item{proxy_pwd}{Password for proxy if needed.}
}
\value{
a \code{data.table}
}
\description{
Get hydraulique data drom eco2mix
}
\examples{
\dontrun{

fil_eau <- get_hydraulique_fil_de_l_eau_eclusee(
  user = "NNI", proxy_pwd = "PASSWORD"
)

}
}
