% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rte-api.R
\name{get_ntc}
\alias{get_ntc}
\title{Retrieve NTC data via RTE data API}
\usage{
get_ntc(token, type = c("ANNUAL", "MONTHLY", "WEEKLY", "D-1", "CURTAILED"),
  start_date = NULL, end_date = NULL, country_eic_code = NULL,
  user = NULL, proxy_pwd = NULL)
}
\arguments{
\item{token}{Token obtained with \code{\link{get_token}}.}

\item{type}{NTC due type, mandatory, one or several
between 'ANNUAL', 'MONTHLY', 'WEEKLY', 'D-1', 'CURTAILED'.}

\item{start_date}{Optional, starting date to filter results,
if used, \code{end_date} must be set as well.}

\item{end_date}{Optional, ending date to filter results.}

\item{country_eic_code}{Country code.}

\item{user}{Username (NNI) for proxy if needeed.}

\item{proxy_pwd}{Password for proxy if needeed.}
}
\value{
a \code{data.table}.
}
\description{
Retrieve NTC data via RTE data API
}
\examples{
\dontrun{

# First you need a token
id_client <- "XXXXX-XXXXX-XXXXX-XXXXX-XXXXX"
id_secret <- "XXXXX-XXXXX-XXXXX-XXXXX-XXXXX"
token <- get_token(
  key = list(id_client = id_client, id_secret = id_secret)
)

# Then you can retrieve NTC data
ntc <- get_ntc(token = token, type = c("ANNUAL", "MONTHLY"))

}
}
