% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_previous}
\alias{get_previous}
\title{Get previous day from a date}
\usage{
get_previous(what = "samedi", date = Sys.Date())
}
\arguments{
\item{what}{Name of the day to retrieve previous the specified date}

\item{date}{A date, by default the current date}
}
\value{
a Date
}
\description{
Get previous day from a date
}
\examples{

# Previous friday
get_previous(what = "vendredi")

# You can abreviate
get_previous(what = "ven")

# Previous saturday before previous friday
get_previous(what = "samedi", date = get_previous(what = "vendredi"))

# Alternatively you can pass a number between 1-7 where 1 is Monday
get_previous(what = 7)

}
