% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rte-api.R
\name{get_proxy_info}
\alias{get_proxy_info}
\title{Get proxy information}
\usage{
get_proxy_info(user = NULL, proxy_pwd = NULL)
}
\arguments{
\item{user}{Optionnal, username (NNI) for proxy.}

\item{proxy_pwd}{Optionnal, password for proxy.}
}
\value{
a list with username and password for proxy
}
\description{
Get proxy information
}
\examples{
\dontrun{

# If you have set your proxy informations, 
# the list should contain them
get_proxy_info()

}
}
