% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mono.R
\name{process_data_mono}
\alias{process_data_mono}
\title{Process Monotone Data}
\usage{
process_data_mono(start = "2016-11-05", date = "2016-11-10 17:00:00",
  area = "fr", nb_MC = 2040, opts = antaresRead::simOptions())
}
\arguments{
\item{start}{Starting date}

\item{date}{Date desired}

\item{area}{Area}

\item{nb_MC}{Monte-Carlo year to look for}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
a list of \code{data.table}
}
\description{
THIS FUNCTION IS DEPRECATED, PLEASE USE \code{\link{compute_mono}}.
}
