% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-capa-hydro.R
\name{read_capa_hydro}
\alias{read_capa_hydro}
\title{Read Hydraulic Capacity}
\usage{
read_capa_hydro(path)
}
\arguments{
\item{path}{Path to the CSV file. If the argument is not specified, a dialog
box will open to select a directory.}
}
\value{
A \code{numeric}.
}
\description{
Read Hydraulic Capacity
}
\examples{
\dontrun{

read_capa_hydro(path = "capa_hydro/PH_RTEHEB_2018_S17_2.csv")
read_capa_hydro(path = "capa_hydro/PH_RTEHEB_2018_S09_5.csv")

}

}
