% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-cnes.R
\name{read_cnes}
\alias{read_cnes}
\title{Read forecast from CNES}
\usage{
read_cnes(path)
}
\arguments{
\item{path}{Path to the file(s) to read, if path is a directory,
all files starting with \code{prev} are read.}
}
\value{
a \code{data.table}
}
\description{
Read forecast from CNES
}
\examples{
\dontrun{

library(antaresWeeklyMargin)

# read only one file
prev1 <- read_cnes("Prev_hebdo_CNES/prev1")
prev1

# read several files
prev12 <- read_cnes(c(
  "Prev_hebdo_CNES/prev1",
  "Prev_hebdo_CNES/prev2"
))
prev12

# read the whole directory
prev_all <- read_cnes("Prev_hebdo_CNES/")
prev_all

}
}
