% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco2mix.R
\name{read_eco2mix}
\alias{read_eco2mix}
\title{Read eco2mix data}
\usage{
read_eco2mix(path)
}
\arguments{
\item{path}{Path to file.}
}
\description{
Read eco2mix data
}
\examples{
\dontrun{

eco2mix <- read_eco2mix("eCO2mix_RTE_En-cours-Consolide.xls")

}
}
