% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-forecast-power.R
\name{read_forecast_power}
\alias{read_forecast_power}
\title{Read Power Forecasts}
\usage{
read_forecast_power(path)
}
\arguments{
\item{path}{path to a directory containing forecast files or a single file.}
}
\value{
a\code{data.table}
}
\description{
Read Power Forecasts
}
\examples{
\dontrun{

fore_pow <- read_forecast_power(path = "inputs/directory/")

}
}
