% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-hydro-xml.r
\name{read_hydro_usine}
\alias{read_hydro_usine}
\title{Read hydro usine XML file}
\usage{
read_hydro_usine(path)
}
\arguments{
\item{path}{Path to the XML file. Can be left blank, designate an XML
file or directory containing XML files. If the path points to a directory,
the most recent XML file is read. If the argument is not specified, a dialog
box will open to select a directory.}
}
\value{
a \code{data.table}
}
\description{
Read hydro usine XML file
}
\examples{
\dontrun{

usine <- read_hydro_usine("inputs/hydros/")

}
}
