% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-meteologica.R
\name{read_meteologica}
\alias{read_meteologica}
\title{Read Meteologica file}
\usage{
read_meteologica(path, country = NULL)
}
\arguments{
\item{path}{path to a directory containing forecast files or a single file.}

\item{country}{filter files to read onlyone country}
}
\value{
a \code{data.table}
}
\description{
Read Meteologica file
}
\examples{
\dontrun{

# one file
one_file <- read_meteologica(
  path = "20180214/PhotovoltaicPower/country-type-code-code_000000_00.csv"
)

# filter with a country
country <- read_meteologica(path = "20180214/PhotovoltaicPower/", country = "France")


# all files in subdir PhotovoltaicPower/
all_pp <- read_meteologica(path = "20180214/PhotovoltaicPower/")
all_pp[, .N, by = list(type, country)]


# all files in subdir 20180214/
all <- read_meteologica(path = "20180214/")
all[, .N, by = list(type, country)]

}
}
