% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-meteologica2.R
\name{read_meteologica2}
\alias{read_meteologica2}
\title{Read Meteologica file}
\usage{
read_meteologica2(path, country = NULL)
}
\arguments{
\item{path}{path to a directory containing forecast files or a single file.}

\item{country}{filter files to read onlyone country}
}
\value{
a \code{data.table}
}
\description{
Read Meteologica file
}
\examples{
\dontrun{

# TODO

}
}
