% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-ntc.R
\name{read_ntc}
\alias{read_ntc}
\title{read NTC files}
\usage{
read_ntc(path)
}
\arguments{
\item{path}{Path to a directory containing CSV files.}
}
\value{
a \code{data.table}
}
\description{
read NTC files
}
\examples{
\dontrun{

ntc <- read_ntc(path = "inputs/ntc_transparency")

}
}
