% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-suppliers.R
\name{read_planning}
\alias{read_planning}
\title{Read planning from suppliers}
\usage{
read_planning(path)
}
\arguments{
\item{path}{Path to the Excel file(s). Can be left blank, designate an Excel
file or directory containing Excel files. If the path points to a directory,
all Excel file are read. If the argument is not specified, a dialog
box will open to select a directory.}
}
\value{
a \code{data.table}
}
\description{
Read planning from suppliers
}
