% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-planning-edf.R
\name{read_planning_edf}
\alias{read_planning_edf}
\title{Read EDF thermal file}
\usage{
read_planning_edf(path)
}
\arguments{
\item{path}{Path to the Excel file. Can be left blank, designate an Excel
file or directory containing Excel files. If the path points to a directory,
the most recent Excel file is read. If the argument is not specified, a dialog
box will open to select a directory.}
}
\value{
a \code{data.table}
}
\description{
Read EDF thermal file
}
