% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-suppliers.R
\name{read_planning_gdfsuez}
\alias{read_planning_gdfsuez}
\title{Read planning from GDF-SUEZ}
\usage{
read_planning_gdfsuez(path)
}
\arguments{
\item{path}{Path to the XML file. Can be left blank, designate an XML
file or directory containing XML files. If the path points to a directory,
the most recent XML file is read. If the argument is not specified, a dialog
box will open to select a directory.}
}
\value{
a \code{data.table}
}
\description{
Read planning from GDF-SUEZ
}
