% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-suppliers.R
\name{read_planning_trc}
\alias{read_planning_trc}
\title{Read planning from Total Raffinage Chimie}
\usage{
read_planning_trc(path)
}
\arguments{
\item{path}{Path to the Excel file. Can be left blank, designate an Excel
file or directory containing Excel files. If the path points to a directory,
the most recent Excel file is read. If the argument is not specified, a dialog
box will open to select a directory.}
}
\value{
a \code{data.table}
}
\description{
Read planning from Total Raffinage Chimie
}
