% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-pump.R
\name{read_pump}
\alias{read_pump}
\title{Read hydraulic pump files}
\usage{
read_pump(path)
}
\arguments{
\item{path}{Path to a directory containing Excel files, both PE & PE file must exist.}
}
\value{
a \code{data.table}
}
\description{
Read hydraulic pump files
}
