% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-thermal-xml.r
\name{read_thermal_tranche}
\alias{read_thermal_tranche}
\title{Read thermal XML file}
\usage{
read_thermal_tranche(path)
}
\arguments{
\item{path}{Path to the XML file. Can be left blank, designate an XML
file or directory containing XML files. If the path points to a directory,
the most recent XML file is read. If the argument is not specified, a dialog
box will open to select a directory.}
}
\value{
a \code{data.table}
}
\description{
Read thermal XML file
}
\examples{
\dontrun{

thermal <- read_thermal_tranche("inputs/thermals/")

}
}
