% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEnergy.R
\name{readEnergy}
\alias{readEnergy}
\title{Read energy data for an area}
\usage{
readEnergy(area, timeStep = "daily", opts = antaresRead::simOptions())
}
\arguments{
\item{area}{An 'antares' area.}

\item{timeStep}{Resolution of the data to import: daily
(default, values are divided by number of days), weekly, monthly (original data).}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\description{
Read energy data for an area
}
\examples{
\dontrun{
library(antaresWeeklyMargin)
library(antaresRead)

# set path to your simulation
opts <- setSimulationPath(path = "test_case/", simulation = "input")

weekly <- readEnergy(area = "fr", timeStep = "weekly", opts = opts)
weekly

daily <- readEnergy(area = "fr", timeStep = "daily", opts = opts)
daily

monthly <- readEnergy(area = "fr", timeStep = "monthly", opts = opts)
monthly
}
}
