% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_wm.R
\name{run_ts_wm}
\alias{run_ts_wm}
\title{Run Time-series Generator for Weekly margin Simulation}
\usage{
run_ts_wm(path_solver = getOption("antares.solver"),
  opts = antaresRead::simOptions())
}
\arguments{
\item{path_solver}{Character containing the Antares Solver path.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}.}
}
\description{
Run Time-series Generator for Weekly margin Simulation
}
