% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_wm.R
\name{setup-wm}
\alias{setup-wm}
\alias{sim_wm}
\alias{path_sim_wm}
\alias{force_path}
\title{Setup a WM simulation before running Antares}
\usage{
sim_wm(date_prev, start_prev_hebdo, path_inputs = path_sim_wm(),
  n_mcyears = 2040, type_load = "prevu",
  load_offset_options = offset_opts(), dispo_pump = c(3520, 3520, 3520,
  3520, 3520, 3520, 3520), simulation_source = NULL,
  simulation_dest = NULL, copy_study = TRUE,
  opts = antaresRead::simOptions())

path_sim_wm(path_dir = "inputs", meteologica = "meteologica",
  cnes = "cnes", planning = "planning", forfait_oa = "forfait_oa",
  ntc = "ntc", ntc_tp = "ntc_tp", capa_hydro = "capa_hydro",
  hydro = "hydro")

force_path(...)
}
\arguments{
\item{date_prev}{Date of simulation.}

\item{start_prev_hebdo}{Date of forecasts, with format \code{\%Y-\%m-\%d}.}

\item{path_inputs}{A \code{list} with path to inputs directories, obtained with \code{path_sim_wm}.}

\item{n_mcyears}{Number of MC years in the study, default to \code{2040}.}

\item{type_load}{Forecast to use \code{"prevu"}, \code{"premis"} or \code{"offset"}.}

\item{load_offset_options}{Peak and off-peak if \code{type_load = "offset"}, see \code{\link{offset_opts}}.}

\item{dispo_pump}{Pumpage availability.}

\item{simulation_source}{Path to source simulation for creating Hydro for other areas.
If provided a copy of this simulation will be performed.}

\item{simulation_dest}{Name of the directory where to copy the study. Warning: content of directory will be deleted!}

\item{copy_study}{Make a copy of the whole study before modifying Antares inputs.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}

\item{path_dir}{Main directory where inputs are.}

\item{meteologica}{Sub-directory containing meteologica files.}

\item{cnes}{Sub-directory containing CNES files.}

\item{planning}{Sub-directory containing planning files.}

\item{forfait_oa}{Sub-directory containing forfait OA files.}

\item{ntc}{Sub-directory containing NTC files.}

\item{ntc_tp}{Sub-directory containing NTC transparency files.}

\item{capa_hydro}{Sub-directory containing Hydraulic capacity transparency files.}

\item{hydro}{Sub-directory containing hydro files.}

\item{...}{Character vectors indicating path to a directory,
use this to specify a path outside the \code{"path_dir"} directory.}
}
\value{
For \code{path_sim_wm} a named \code{list}.
}
\description{
Setup a WM simulation before running Antares

Set path to inputs for WM simulation
}
\examples{
\dontrun{

library(antaresRead)
library(antaresWeeklyMargin)

opts <- setSimulationPath("path/to/original/study")

sim_wm(
  date_prev = "2018-06-21",
  start_prev_hebdo = "2018-06-23", 
  simulation_dest = "test_case_WM",
  type_load = "prevu",
  path_inputs = path_sim_wm(  # Path to input data
    path_dir = "../inputs", 
    meteologica = "meteologica", 
    cnes = "cnes",           
    planning = "planning", 
    forfait_oa = "forfait_oa",  
    ntc = "ntc",         
    ntc_tp = "ntc_tp",    
    capa_hydro = "capa_hydro",  
    hydro = "hydro"   
  ), 
  opts = opts
)

}
}
