% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eco2mix.R
\name{setup_eco2mix}
\alias{setup_eco2mix}
\title{Setup Eco2mix data for Weekly Margin Simulation}
\usage{
setup_eco2mix(path_input)
}
\arguments{
\item{path_input}{Path where eco2mix data are or will downloaded.}
}
\description{
Setup Eco2mix data for Weekly Margin Simulation
}
