% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rte-api.R
\name{setupProxy}
\alias{setupProxy}
\title{Setup proxy credentials}
\usage{
setupProxy(user, proxy_pwd)
}
\arguments{
\item{user}{Username (NNI) for proxy.}

\item{proxy_pwd}{Password for proxy.}
}
\description{
Setup proxy credentials
}
\note{
You'll need to restart your R session for change to be effective.
}
\examples{
\dontrun{

setupProxy("MYNNI", "MY_PASSWORD")
# restart your R session
# info should appear in the list
get_proxy_info()

}
}
