% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_markdown.R
\name{create_markdown}
\alias{create_markdown}
\title{Create directory with SMTA markdown}
\usage{
create_markdown(path, format = c("report", "slides"))
}
\arguments{
\item{path}{Path to a directory, if directory doesn't exist, it will be created.}

\item{format}{Format of markdown output : report or slides}
}
\description{
Create directory with SMTA markdown
}
\examples{
\dontrun{

create_markdown("SMTA_rapport", "report")

}
}
