% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_markdown.R
\name{create_markdown}
\alias{create_markdown}
\title{Create directory with SMTA markdown}
\usage{
create_markdown(path, format = c("report", "slides"), date_start,
  date_study, week, n_scenario, year_mc,
  opts = antaresRead::simOptions())
}
\arguments{
\item{path}{Path to a directory, if directory doesn't exist, it will be created.}

\item{format}{Format of markdown output : report or slides.}

\item{date_start}{Study start date.}

\item{date_study}{Date for margins computation.}

\item{week}{Number of the week studied.}

\item{n_scenario}{Number of scenario used in analysis.}

\item{year_mc}{Number of the Monte-Carlo year to focus on.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\description{
Create directory with SMTA markdown
}
\examples{
\dontrun{

create_markdown("SMTA_rapport", "report")

}
}
