% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wm_cluster.R
\name{create_wm_cluster}
\alias{create_wm_cluster}
\title{Create clusters for Weekly Margins simulation}
\usage{
create_wm_cluster(data, start = NULL, rm_prev_clus = TRUE,
  sort_other_clus = TRUE, opts = antaresRead::simOptions())
}
\arguments{
\item{data}{a \code{data.table} obtained from \code{\link{read_planning}}.}

\item{start}{If specified, data will be filtered from given date to 7 days after.}

\item{rm_prev_clus}{Remove previous clusters before creating new ones.}

\item{sort_other_clus}{Reorder rows of other clusters data.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\description{
Create clusters for Weekly Margins simulation
}
\examples{
\dontrun{

# set path to your simulation
opts <- setSimulationPath(path = "path/", simulation = "input")

# Read data from suppliers
plannings <- read_planning()

# create clusters
clusters_crea <- create_wm_cluster(plannings, opts)

}
}
