% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wm_hydro_fr.R
\name{create_wm_hydro_fr}
\alias{create_wm_hydro_fr}
\title{Create FR Hydro setup for Weekly Margins simulation}
\usage{
create_wm_hydro_fr(path_capa_hydro, path_hydro, start, fictive_area = "lac",
  dispo_pump = c(3200, 3200, 3020, 2860, 3020, 3180, 3180),
  opts = antaresRead::simOptions())
}
\arguments{
\item{path_capa_hydro}{Path to hydro capacities CSV file.}

\item{path_hydro}{Path to hydro usine XML file.}

\item{start}{Starting date of simulation.}

\item{fictive_area}{Name of the fictive area to create, default is \code{"lac"}.}

\item{dispo_pump}{Pumpage availability.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\description{
Create FR Hydro setup for Weekly Margins simulation
}
\examples{
\dontrun{

# todo

}
}
