% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wm_misc.R
\name{create_wm_misc}
\alias{create_wm_misc}
\title{Create Misc-gen for Weekly Margins simulation}
\usage{
create_wm_misc(data, start = NULL, opts = antaresRead::simOptions())
}
\arguments{
\item{data}{a \code{data.table} obtained from \code{\link{read_forfait_oa}}.}

\item{start}{If specified, data will be filtered from given date to 7 days after.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\description{
Create Misc-gen for Weekly Margins simulation
}
\examples{
\dontrun{

# todo

}
}
