% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wm_ror.R
\name{create_wm_ror}
\alias{create_wm_ror}
\title{Create ROR for Weekly Margins simulation}
\usage{
create_wm_ror(data, start = NULL, startday = "samedi",
  sort_areas = TRUE, opts = antaresRead::simOptions())
}
\arguments{
\item{data}{a \code{data.table} obtained from \code{\link{read_forfait_oa}}.}

\item{start}{Starting day of the simulation, data between previous \code{startday}
and next \code{startday}-1, by default between \code{samedi} and \code{vendredi}.}

\item{startday}{Day of week to start simulation.}

\item{sort_areas}{Reorder other ROR data to match the desired week.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\description{
Create ROR for Weekly Margins simulation
}
\examples{
\dontrun{

# set path to your simulation
opts <- setSimulationPath(path = "path/to/simulation/", simulation = "input")

# Read OA files
oa <- read_forfait_oa(path = "path/to/hydro_forfait/")

# Create ROR series in Antares
create_wm_ror(data = oa, start = "2018-01-04")

}
}
