% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-margins.R
\name{draw_margins}
\alias{draw_margins}
\title{Draw remaining capacity line chart}
\usage{
draw_margins(data_margin, area = "fr", num_week = NULL)
}
\arguments{
\item{data_margin}{a \code{data.table}, first column must be the datetime,
the others results of Monte-Carlo simulations.}

\item{area}{Name of the area, used in chart title.}

\item{num_week}{Week number to be displayed in chart title.}
}
\value{
a \code{dygraphs} htmlwidget.
}
\description{
Draw remaining capacity line chart
}
\examples{
\dontrun{

library(antaresRead)
library(antaresWeeklyMargin)

setSimulationPath("path/to/sim")


up <- compute_margins(
  date = "2018-02-28", 
  area = "fr", 
  mcYears = 1:100,
  margin = "upward"
)

# Marges seules
draw_margins(data_margin = up$margin_area_solo, area = "fr")

# Marges interco
draw_margins(data_margin = up$margin_area_inter, area = "fr")

}
}
