% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins_quantiles.R
\name{ft_margins_quantiles}
\alias{ft_margins_quantiles}
\title{Flextable format for margins quantiles}
\usage{
ft_margins_quantiles(margin, layout = c("horizontal", "vertical"),
  language = c("fr", "en"))
}
\arguments{
\item{margin}{Output from \code{\link{margins_quantiles}}.}

\item{layout}{Vertical or horizontal table.}

\item{language}{Language to use : \code{fr} or \code{en}.}
}
\value{
a \code{flextable} object
}
\description{
Flextable format for margins quantiles
}
\examples{
\dontrun{

res <- margin_quantile(marges_inter_fr)
ft_margins_quantiles(res)

}
}
