% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rte-api.R
\name{get_eco2mix}
\alias{get_eco2mix}
\title{Get data from eco2mix via Open-RTE API}
\usage{
get_eco2mix(from = NULL, to = NULL, user = NULL, proxy_pwd = NULL)
}
\arguments{
\item{from}{date from which to retrieve data.}

\item{to}{date until which to recover data.}

\item{user}{Username (NNI) for proxy if needeed.}

\item{proxy_pwd}{Password for proxy if needeed.}
}
\value{
a \code{data.table}
}
\description{
Get data from eco2mix via Open-RTE API
}
\examples{
\dontrun{

eco2mix <- get_eco2mix(
  from = "2018-01-06", 
  to = "2018-01-12", 
  user = "NNI",         # needeed if no internet connection open
  proxy_pwd = "PASSWORD"
)

}
}
