% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins_quantiles.R
\name{margins_quantiles}
\alias{margins_quantiles}
\title{Compute margins quantiles for a couple of hours}
\usage{
margins_quantiles(margin, hours = c("09h", "19h"), type = c("upward",
  "downward"))
}
\arguments{
\item{margin}{a \code{data.table} containing margins.}

\item{hours}{Couple of hours.}

\item{type}{Type of margin.}
}
\value{
a \code{data.table}
}
\description{
Compute margins quantiles for a couple of hours
}
\examples{
\dontrun{

margin_quantile(marges_inter_fr)

}

}
