% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins_quantiles.R
\name{margins_quantiles}
\alias{margins_quantiles}
\title{Compute margins' quantiles}
\usage{
margins_quantiles(margin)
}
\arguments{
\item{margin}{a \code{data.table} containing margins.}
}
\value{
a \code{data.table}
}
\description{
Compute margins' quantiles
}
\examples{
\dontrun{

margin_quantile(marges_inter_fr)

}

}
