% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-forfait.R
\name{read_forfait_oa}
\alias{read_forfait_oa}
\title{Read EDF forfait file}
\usage{
read_forfait_oa(path)
}
\arguments{
\item{path}{Path to the XML file. Can be left blank, designate an XML
file or directory containing XML files. If the path points to a directory,
the most recent XML file is read. If the argument is not specified, a dialog
box will open to select a directory.}
}
\value{
a \code{data.table}
}
\description{
Read EDF forfait file
}
\examples{
\dontrun{

# TODO

}
}
