% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_report.R
\name{render_report}
\alias{render_report}
\title{Render Weekly Margin Report}
\usage{
render_report(output_file = "index.html", date_start, date_study, week,
  n_scenario, year_mc, opts = antaresRead::simOptions())
}
\arguments{
\item{output_file}{PAth to output file generated.}

\item{date_start}{Study start date.}

\item{date_study}{Date for margins computation.}

\item{week}{Number of the week studied.}

\item{n_scenario}{Number of scenario used in analysis.}

\item{year_mc}{Number of the Monte-Carlo year to focus on.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\description{
Render Weekly Margin Report
}
