% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_wm.R
\name{setup-wm}
\alias{setup-wm}
\alias{sim_wm}
\alias{path_sim_wm}
\title{Setup a WM simulation before running Antares}
\usage{
sim_wm(date_prev, start_prev_hebdo, path_inputs = path_sim_wm(),
  type_load = "prevu", dispo_pump = c(3520, 3520, 3520, 3520, 3520, 3520,
  3520), opts = antaresRead::simOptions())

path_sim_wm(path_dir = "inputs", meteologica = "meteologica",
  cnes = "cnes", planning = "planning", forfait_oa = "forfait_oa",
  ntc = "ntc", ntc_tp = "ntc_tp", capa_hydro = "capa_hydro",
  hydro = "hydro")
}
\arguments{
\item{date_prev}{Date of simulation.}

\item{start_prev_hebdo}{Date of forecasts, with format \code{\%Y-\%m-\%d}.}

\item{path_inputs}{A \code{list} with path to inputs directories, obtained with \code{path_sim_wm}.}

\item{type_load}{Forecast to use \code{prevu} or \code{premis}.}

\item{dispo_pump}{Pumpage availability.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}

\item{path_dir}{Main directory where inputs are.}

\item{meteologica}{Sub-directory containing meteologica files.}

\item{cnes}{Sub-directory containing CNES files.}

\item{planning}{Sub-directory containing planning files.}

\item{forfait_oa}{Sub-directory containing forfait OA files.}

\item{ntc}{Sub-directory containing NTC files.}

\item{ntc_tp}{Sub-directory containing NTC transparency files.}

\item{capa_hydro}{Sub-directory containing Hydraulic capacity transparency files.}

\item{hydro}{Sub-directory containing hydro files.}
}
\value{
For \code{path_sim_wm} a named \code{list}.
}
\description{
Setup a WM simulation before running Antares

Set path to inputs for WM simulation
}
\examples{
\dontrun{

# todo

}
}
